Withdraw Money V1.02
UserViewDTO only used for the authentication,
Float to Decimal change on account.balance in db, schemas,router,services

+++ app/api/routes/accounts/schemas.py
@@ -1,10 +1,10 @@
 from pydantic import BaseModel
-
+from decimal import Decimal
 
 class AccountViewDTO(BaseModel):
     id: int
     username: str
-    balance: float
+    balance: Decimal
     _is_blocked: bool
 
     @property
@@ -16,7 +16,7 @@ class AccountViewDTO(BaseModel):
         self._is_blocked = bool(value)
 
     @classmethod
-    def from_query_result(cls, id: int, username: str, balance: int, is_blocked: bool):
+    def from_query_result(cls, id: int, username: str, balance: Decimal, is_blocked: bool):
         return cls(id=id, username=username, balance=balance, is_blocked=is_blocked)
